Uses double instead of float in loops

+++ collect_app/src/main/java/org/odk/collect/android/utilities/ColorPickerDialog.java
@@ -112,32 +112,32 @@ public class ColorPickerDialog extends Dialog {
 
             // Initialize the colors of the hue slider bar
             int index = 0;
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 // Red (#f00) to pink (#f0f)
                 hueBarColors[index] = Color.rgb(255, 0, (int) i);
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 // Pink (#f0f) to blue (#00f)
                 hueBarColors[index] = Color.rgb(255 - (int) i, 0, 255);
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 // Blue (#00f) to light blue (#0ff)
                 hueBarColors[index] = Color.rgb(0, (int) i, 255);
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 // Light blue (#0ff) to green (#0f0)
                 hueBarColors[index] = Color.rgb(0, 255, 255 - (int) i);
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 // Green (#0f0) to yellow (#ff0)
                 hueBarColors[index] = Color.rgb((int) i, 255, 0);
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 // Yellow (#ff0) to red (#f00)
                 hueBarColors[index] = Color.rgb(255, 255 - (int) i, 0);
                 index++;
@@ -153,37 +153,37 @@ public class ColorPickerDialog extends Dialog {
         private int getCurrentMainColor() {
             int translatedHue = 255 - (int) (currentHue * 255 / 360);
             int index = 0;
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 if (index == translatedHue) {
                     return Color.rgb(255, 0, (int) i);
                 }
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 if (index == translatedHue) {
                     return Color.rgb(255 - (int) i, 0, 255);
                 }
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 if (index == translatedHue) {
                     return Color.rgb(0, (int) i, 255);
                 }
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 if (index == translatedHue) {
                     return Color.rgb(0, 255, 255 - (int) i);
                 }
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 if (index == translatedHue) {
                     return Color.rgb((int) i, 255, 0);
                 }
                 index++;
             }
-            for (float i = 0; i < 256; i += 256 / 42) {
+            for (double i = 0; i < 256; i += 256 / 42) {
                 if (index == translatedHue) {
                     return Color.rgb(255, 255 - (int) i, 0);
                 }
